requirements.txt
flask
google-generativeai
requests
pymongo